@@ -6,13 +6,16 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  dio:
-  shared_preferences:
+  dio: ^4.0.0
+  flutter_mobx: ^2.0.0
+  flutter_modular: ^3.0.3
+  mobx: ^2.0.1
+  shared_preferences: ^2.0.5
 
 dev_dependencies:
   flutter_test:
