@@ -2,15 +2,15 @@ import 'package:oneonones/common/models/employee.model.dart';
 import 'package:oneonones/common/models/oneonone_compose.model.dart';
 
 class DashboardModel {
-  EmployeeModel employee;
-  List<OneononeComposeModel> oneonones;
+  EmployeeModel? employee;
+  List<OneononeComposeModel>? oneonones;
 
   DashboardModel({
     this.employee,
     this.oneonones,
   });
 
-  DashboardModel.fromJson(Map<String, dynamic> json) {
+  DashboardModel.fromJson(Map<String, dynamic>? json) {
     if (json == null) return;
     employee = EmployeeModel.fromJson(json['employee']);
     oneonones = List<OneononeComposeModel>.from(json['oneonones'].map((j) => OneononeComposeModel.fromJson(j)));
