@@ -1,9 +1,9 @@
 class ErrorModel {
-  List<String> errors;
+  List<String>? errors;
 
   ErrorModel({this.errors});
 
-  ErrorModel.fromJson(Map<String, dynamic> json) {
+  ErrorModel.fromJson(Map<String, dynamic>? json) {
     if (json == null) return;
     errors = json['errors'].cast<String>();
   }
