@@ -1,10 +1,10 @@
 import 'package:oneonones/common/models/employee.model.dart';
 
 class OneononeHistoricalModel {
-  EmployeeModel leader;
-  EmployeeModel led;
-  DateTime occurrence;
-  String commentary;
+  EmployeeModel? leader;
+  EmployeeModel? led;
+  DateTime? occurrence;
+  String? commentary;
 
   OneononeHistoricalModel({
     this.leader,
@@ -13,7 +13,7 @@ class OneononeHistoricalModel {
     this.commentary,
   });
 
-  OneononeHistoricalModel.fromJson(Map<String, dynamic> json) {
+  OneononeHistoricalModel.fromJson(Map<String, dynamic>? json) {
     if (json == null) return;
     leader = EmployeeModel.fromJson(json['leader']);
     led = EmployeeModel.fromJson(json['led']);
