@@ -1,8 +1,8 @@
 class OneononeHistoricalInputModel {
-  String leaderEmail;
-  String ledEmail;
-  DateTime occurrence;
-  String commentary;
+  String? leaderEmail;
+  String? ledEmail;
+  DateTime? occurrence;
+  String? commentary;
 
   OneononeHistoricalInputModel({
     this.leaderEmail,
@@ -15,7 +15,7 @@ class OneononeHistoricalInputModel {
     return {
       'leaderEmail': leaderEmail,
       'ledEmail': ledEmail,
-      'occurrence': occurrence.toIso8601String().substring(0, 10),
+      'occurrence': occurrence!.toIso8601String().substring(0, 10),
       'commentary': commentary,
     };
   }
